Initial version of scheduling support.
This version doesn't actually work yet - I'm working through a support
ticket with Constant Contact.  The primary problem is that saving a
schedule causes a 403 Forbidden error (when using OAuth2).  I can work
around that by adding the OAuth token to the URL as a query strings
(instead of as an Authorization header), but I'm waiting to see if
Constant Contact can come up with something else.