Pass headers when paging thru results from CC
It's not well documented, but if a call to the CC API like get mailing lists
has many results, CC only returns the first 50.  Pagination is handled in
atom_format.rb.  With the addition/hack of the oauth2 support that relies on
headers, those headers have to be added to the subsequent calls that page
through the results from CC, otherwise 401 Unauthorized errors occurr.